requirements.txt
mkdocs-material 
mkdocs-macros-plugin
mkdocs-include-markdown-plugin
mkdocs-awesome-pages-plugin
pyyaml